Test runs. Fixed extraneous parenthesis in block_header init, need to figure out issue with page content adresses in main

+++ includes/data_structs.h
@@ -30,7 +30,7 @@ typedef struct t_page
     struct t_page*     next;
     s_block*           block_head;
     e_zone_type        type;
-    size_t             free_space; // Have some way to 'preview' the page so that we do not always need to scan it??
+    long long          free_space; // Have some way to 'preview' the page so that we do not always need to scan it??
 } s_page;
 
 # define PAYLOAD_HEADER(block_ptr) ((char *) (block_ptr) - sizeof(s_block));
